Move Header component out of the Home Hero component
