+++ src/components/index.js
@@ -3,3 +3,4 @@ export { ErrorView } from '@/components/ErrorView';
 export { TabBarIcon } from '@/components/TabBarIcon';
 export { TextField } from '@/components/TextField';
 export { Hero } from '@/components/Hero';
+export { Header } from '@/components/Header';
